RK_KERNEL_CONFIG_TYPE: add WIP ideas
Here is a work-in-progress commit to show the gist of what I'm hoping will
evolve into a more sophisticated way of allowing BSP maintainers and users to
configure their kernel.

At this point I have a switch, RK_KERNEL_CONFIG_TYPE, that allows a user to
specify either using the generic in-kernel defconfig or a more specific
in-layer defconfig. Currently only rock-pi-4b has support for "inlayer", but
this is just to show how it might work.

Ideally overlaid on top of a more basic kernel config: selecting a machine
specifies a rockchip (which selects specific set of config fragments) and
selects an architecture (which selects more specific config fragments) and
also implies a board (which selects even more specific config fragments). Thus
building up a specific kernel config for this specific SoC, CPU, and board.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>